refactor(stylelint): renames rules and corrected messages based on cx review
