+++ packages/stylelint-plugin-slds/src/rules/index.ts
@@ -12,8 +12,8 @@ import enforceComponentHookNamingConvention from './enforce-component-hook-namin
 import reduceAnnotations from './reduce-annotations';
 import { noHardcodedValuesSlds, noHardcodedValuesSldsPlus } from './no-hardcoded-value';
 import noSldsVarWithoutFallback from './no-slds-var-without-fallback';
-import noSldsNamespace from './no-slds-namespace';
-import noUnsupportedVarFallback from './no-unsupported-var-fallback';
+import noSldsNamespaceForCustomHooks from './no-slds-namespace-for-custom-hooks';
+import noSldshookFallbackForLwcToken from './no-sldshook-fallback-for-lwctoken';
 
 export default [
   enforceSdsToSldsHooks,
@@ -31,6 +31,6 @@ export default [
   noImportantTag,
   reduceAnnotations,
   noSldsVarWithoutFallback,
-  noSldsNamespace,
-  noUnsupportedVarFallback
+  noSldsNamespaceForCustomHooks,
+  noSldshookFallbackForLwcToken
 ];
