+++ packages/stylelint-plugin-slds/src/rules/no-sldshook-fallback-for-lwctoken/index.ts
@@ -4,7 +4,7 @@ import ruleMetadata from '../../utils/rulesMetadata';
 import replacePlaceholders from '../../utils/util';
 import metadata from '@salesforce-ux/sds-metadata';
 import { isTargetProperty } from '../../utils/prop-utills';
-import { forEachVarFunction } from '../../utils/decl-utils';
+import { forEachVarFunction, getFallbackToken, getVarToken } from '../../utils/decl-utils';
 import valueParser from 'postcss-value-parser';
 
 const sldsPlusStylingHooks = metadata.sldsPlusStylingHooks;
@@ -17,48 +17,47 @@ const allSldsHooks = [...sldsPlusStylingHooks.global, ...sldsPlusStylingHooks.co
 
 const { utils, createPlugin }: typeof stylelint = stylelint;
 
-const ruleName: string = 'slds/no-slds-namespace';
+const ruleName: string = 'slds/no-sldshook-fallback-for-lwctoken';
 
 const { severityLevel = 'error', warningMsg = '', errorMsg = '', ruleDesc = 'No description provided' } = ruleMetadata(ruleName) || {};
 
-const toSldsToken = (sdsToken: string) => sdsToken.replace('--sds-', '--slds-')
+const toSldsToken = (sdsToken: string='') => (sdsToken || '').replace('--sds-', '--slds-')
 
 const messages = stylelint.utils.ruleMessages(ruleName, {
-    expected: (token: string) => {
-        const tokenWithoutNamespace = token.replace('--slds-', '').replace('--sds-', '');
-        return replacePlaceholders(warningMsg, { token, tokenWithoutNamespace })
+    expected: (lwcToken: string, sldsToken: string) => {
+        return replacePlaceholders(warningMsg, { lwcToken, sldsToken })
     },
 });
 
-function shouldIgnoreDetection(sdsToken: string) {
-    // Ignore if entry found in the list or not starts with reserved namespace
-    if(sdsToken.startsWith('--sds-') || sdsToken.startsWith('--slds-')){
-        return allSldsHooks.includes(toSldsToken(sdsToken))
-    }
-    return true;
+function hasUnsupportedFallback(lwcToken: string, sldsToken: string): boolean {
+    const safeSldsToken = toSldsToken(sldsToken);
+    return lwcToken && safeSldsToken 
+    && lwcToken.startsWith('--lwc-') 
+    && safeSldsToken.startsWith('--slds-') 
+    && allSldsHooks.includes(safeSldsToken);
 }
 
 /**
  * 
  * Example:
  *  .THIS  .demo {
- *    border: 1px solid var(--slds-my-own-token));
+ *    color: var(--lwc-color-background-1, var(--sds-g-color-background-1));
  *  }
  * 
  */
 function detectRightSide(decl: Declaration, basicReportProps: Partial<stylelint.Problem>) {
 
     forEachVarFunction(decl, (node: valueParser.FunctionNode, startOffset: number) => {
-        const tokenNode = node.nodes[0];
-        const oldValue = tokenNode.value;
-        if (shouldIgnoreDetection(oldValue)) {
-            // Ignore if entry not found in the list or the token is marked to use further
+        const lwcToken = getVarToken(node);
+        const sldsToken = getFallbackToken(node);
+
+        if(!hasUnsupportedFallback(lwcToken, sldsToken)){
             return;
         }
 
-        const index = startOffset + tokenNode.sourceIndex;
-        const endIndex = startOffset + tokenNode.sourceEndIndex;
-        const message = messages.expected(oldValue);
+        const index = startOffset + node.sourceIndex;
+        const endIndex = startOffset + node.sourceEndIndex;
+        const message = messages.expected(lwcToken, sldsToken);
 
         utils.report(<stylelint.Problem>{
             message,
@@ -66,35 +65,7 @@ function detectRightSide(decl: Declaration, basicReportProps: Partial<stylelint.
             endIndex,
             ...basicReportProps
         });
-    });
-}
-
-/**
- * 
- * Example:
- *  .THIS  .demo {
- *    --slds-my-own-token: 50%;
- *  }
- * 
- */
-function detectLeftSide(decl: Declaration, basicReportProps: Partial<stylelint.Problem>) {
-    // Usage on left side
-    const { prop } = decl;
-    if (shouldIgnoreDetection(prop)) {
-        // Ignore if entry not found in the list or the token is marked to use further
-        return;
-    }
-    const startIndex = decl.toString().indexOf(prop);
-    const endIndex = startIndex + prop.length;
-
-    const message = messages.expected(prop);
-
-    utils.report(<stylelint.Problem>{
-        message,
-        index: startIndex,
-        endIndex,
-        ...basicReportProps
-    });
+    }, false);
 }
 
 
@@ -115,7 +86,6 @@ const ruleFunction: Partial<stylelint.Rule> = (primaryOptions: boolean, { severi
             };
 
             detectRightSide(decl, basicReportProps);
-            detectLeftSide(decl, basicReportProps);
         });
     };
 };
