+++ packages/stylelint-plugin-slds/tests/rules/no-slds-namespace-for-custom-hooks/no-slds-namespace-for-custom-hooks.spec.ts
@@ -3,27 +3,27 @@ import sldsPlugin from '../../../src/index';
 
 const { lint }: typeof stylelint = stylelint;
 
-describe('no-slds-namespace', () => {
+describe('no-slds-namespace-for-custom-hooks', () => {
   const testCases = [
     {
       message:
-        'Using slds namespace for --slds-my-own-token isn\'t supported. To differentiate SLDS and custom tokens, create a token in your namespace. Example: --myapp-my-own-token',
+        'Using the --slds namespace for --slds-my-own-token isn\'t supported. Create the custom styling hook in your namespace. Example: --myapp-my-own-token',
       code: `
         .example {
           --slds-my-own-token: #fff;
         }
       `,
-      ruleName: 'slds/no-slds-namespace'
+      ruleName: 'slds/no-slds-namespace-for-custom-hooks'
     },
     {
       message:
-        'Using slds namespace for --slds-custom-color isn\'t supported. To differentiate SLDS and custom tokens, create a token in your namespace. Example: --myapp-custom-color',
+        'Using the --slds namespace for --slds-custom-color isn\'t supported. Create the custom styling hook in your namespace. Example: --myapp-custom-color',
       code: `
         .example {
           color: var(--slds-custom-color);
         }
       `,
-      ruleName: 'slds/no-slds-namespace'
+      ruleName: 'slds/no-slds-namespace-for-custom-hooks'
     },
     {
       message: null,
@@ -32,7 +32,7 @@ describe('no-slds-namespace', () => {
           --myapp-valid-token: #fff;
         }
       `,
-      ruleName: 'slds/no-slds-namespace'
+      ruleName: 'slds/no-slds-namespace-for-custom-hooks'
     },
     {
       message: null,
@@ -41,7 +41,7 @@ describe('no-slds-namespace', () => {
           color: var(--myapp-valid-token);
         }
       `,
-      ruleName: 'slds/no-slds-namespace'
+      ruleName: 'slds/no-slds-namespace-for-custom-hooks'
     },
     {
       message: null,
@@ -50,7 +50,7 @@ describe('no-slds-namespace', () => {
           --slds-c-button-color-background: #fff;
         }
       `,
-      ruleName: 'slds/no-slds-namespace'
+      ruleName: 'slds/no-slds-namespace-for-custom-hooks'
     },
     {
       message: null,
@@ -59,7 +59,7 @@ describe('no-slds-namespace', () => {
           color: var(--slds-c-button-color-background);
         }
       `,
-      ruleName: 'slds/no-slds-namespace'
+      ruleName: 'slds/no-slds-namespace-for-custom-hooks'
     },
   ];
