+++ packages/stylelint-plugin-slds/tests/rules/no-sldshook-fallback-for-lwctoken/no-sldshook-fallback-for-lwctoken.spec.ts
@@ -3,46 +3,46 @@ import sldsPlugin from '../../../src/index';
 
 const { lint }: typeof stylelint = stylelint;
 
-describe('no-unsupported-var-fallback', () => {
+describe('no-sldshook-fallback-for-lwctoken', () => {
   const testCases = [
     {
       message:
-        'Using --slds-g-color-border-1 as fallback value for --lwc-color-background-1 isn\'t supported. For more information, see the Lightning Web Components Developer Guide.',
+        'Remove the --slds-g-color-border-1 styling hook that is used as a fallback value for --lwc-color-background-1.',
       code: `
         .example {
           color: var(--lwc-color-background-1, var(--slds-g-color-border-1));
         }
       `,
-      ruleName: 'slds/no-unsupported-var-fallback'
+      ruleName: 'slds/no-sldshook-fallback-for-lwctoken'
     },
     {
       message:
-        'Using --slds-g-color-border-1 as fallback value for --lwc-color-background-1 isn\'t supported. For more information, see the Lightning Web Components Developer Guide.',
+        'Remove the --slds-g-color-border-1 styling hook that is used as a fallback value for --lwc-color-background-1.',
       code: `
         .example {
           color: var(--lwc-color-background-1, var(--slds-g-color-border-1));
         }
       `,
-      ruleName: 'slds/no-unsupported-var-fallback'
+      ruleName: 'slds/no-sldshook-fallback-for-lwctoken'
     },
     {
       message:
-        'Using --slds-g-color-border-1 as fallback value for --lwc-color-background-1 isn\'t supported. For more information, see the Lightning Web Components Developer Guide.',
+        'Remove the --slds-g-color-border-1 styling hook that is used as a fallback value for --lwc-color-background-1.',
       code: `
         .example {
           color: var(--lwc-color-background-1, var(--slds-g-color-border-1, var(--lwc-color-background-2)));
         }
       `,
-      ruleName: 'slds/no-unsupported-var-fallback'
+      ruleName: 'slds/no-sldshook-fallback-for-lwctoken'
     },
     {
-      message: 'Using --slds-g-color-border-2 as fallback value for --lwc-color-background-1 isn\'t supported. For more information, see the Lightning Web Components Developer Guide.',
+      message: 'Remove the --slds-g-color-border-2 styling hook that is used as a fallback value for --lwc-color-background-1.',
       code: `
         .example {
           color: var(--slds-g-color-border-1, var(--lwc-color-background-1, var(--slds-g-color-border-2)));
         }
       `,
-      ruleName: 'slds/no-unsupported-var-fallback'
+      ruleName: 'slds/no-sldshook-fallback-for-lwctoken'
     },
     {
       message: null,
@@ -51,7 +51,7 @@ describe('no-unsupported-var-fallback', () => {
           color: var(--myapp-color-background-1, var(--slds-g-color-border-1));
         }
       `,
-      ruleName: 'slds/no-unsupported-var-fallback'
+      ruleName: 'slds/no-sldshook-fallback-for-lwctoken'
     },
     {
       message: null,
@@ -60,7 +60,7 @@ describe('no-unsupported-var-fallback', () => {
           color: var(--lwc-color-background-1, var(--myapp-color-border-1));
         }
       `,
-      ruleName: 'slds/no-unsupported-var-fallback'
+      ruleName: 'slds/no-sldshook-fallback-for-lwctoken'
     },
     {
       message: null,
@@ -69,7 +69,7 @@ describe('no-unsupported-var-fallback', () => {
           color: var(--slds-g-color-border-1);
         }
       `,
-      ruleName: 'slds/no-unsupported-var-fallback'
+      ruleName: 'slds/no-sldshook-fallback-for-lwctoken'
     },
     {
       message: null,
@@ -78,7 +78,7 @@ describe('no-unsupported-var-fallback', () => {
           color: var(--lwc-color-background-1);
         }
       `,
-      ruleName: 'slds/no-unsupported-var-fallback'
+      ruleName: 'slds/no-sldshook-fallback-for-lwctoken'
     },
   ];
